refactor: adopt hexagonal architecture
