+++ src/main/java/com/example/notificationservice/adapters/in/web/NotificationController.java
@@ -1,26 +1,25 @@
-package com.example.notificationservice.controller;
+package com.example.notificationservice.adapters.in.web;
 
-import com.example.notificationservice.provider.SmsSubscription;
-import com.example.notificationservice.service.NotificationService;
+import com.example.notificationservice.domain.SmsSubscription;
+import com.example.notificationservice.application.port.in.NotificationUseCase;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
-
 @RestController
 @RequestMapping("/notify")
 public class NotificationController {
-    private final NotificationService notificationService;
+    private final NotificationUseCase notificationUseCase;
 
-    public NotificationController(NotificationService notificationService) {
-        this.notificationService = notificationService;
+    public NotificationController(NotificationUseCase notificationUseCase) {
+        this.notificationUseCase = notificationUseCase;
     }
 
     @PostMapping("/email")
     public ResponseEntity<String> sendEmail(@RequestBody Map<String, String> payload) {
         String to = payload.get("to");
         String message = payload.get("message");
-        notificationService.sendEmail(to, message);
+        notificationUseCase.sendEmail(to, message);
         return ResponseEntity.ok("Email sent");
     }
 
@@ -30,7 +29,7 @@ public class NotificationController {
         String message = payload.get("message");
         String subscriptionValue = payload.get("subscription");
         SmsSubscription subscription = SmsSubscription.valueOf(subscriptionValue.toUpperCase());
-        notificationService.sendSms(subscription, to, message);
+        notificationUseCase.sendSms(subscription, to, message);
         return ResponseEntity.ok("SMS sent");
     }
 }
