+++ src/main/java/com/example/notificationservice/application/port/in/NotificationUseCase.java
@@ -1,0 +1,9 @@
+package com.example.notificationservice.application.port.in;
+
+import com.example.notificationservice.domain.SmsSubscription;
+
+public interface NotificationUseCase {
+    void sendEmail(String to, String message);
+
+    void sendSms(SmsSubscription subscription, String to, String message);
+}
