+++ src/main/java/com/example/notificationservice/domain/NotificationType.java
@@ -1,4 +1,4 @@
-package com.example.notificationservice.provider;
+package com.example.notificationservice.domain;
 
 public enum NotificationType {
     SMS,
