+++ src/main/java/com/example/notificationservice/domain/SmsSubscription.java
@@ -1,4 +1,4 @@
-package com.example.notificationservice.provider;
+package com.example.notificationservice.domain;
 
 public enum SmsSubscription {
     KALEYRA,
